Move LockableObservableListWrapper to internals namespace
