+++ Xamarin.Forms.Platform.WinRT/PickerRenderer.cs
@@ -53,7 +53,7 @@ namespace Xamarin.Forms.Platform.WinRT
 					Control.Loaded += ControlOnLoaded;
 				}
 
-				Control.ItemsSource = ((Picker.LockableObservableListWrapper)Element.Items)._list;
+				Control.ItemsSource = ((LockableObservableListWrapper)Element.Items)._list;
 
 				UpdateTitle();
 				UpdateSelectedIndex();
